+++ stwo_cairo_prover/crates/cairo-air/src/opcodes_air.rs
@@ -104,7 +104,7 @@ impl OpcodeClaim {
         ))
     }
 
-    pub fn accumulate_relation_uses(&self, relation_counts: &mut HashMap<&'static str, u32>) {
+    pub fn accumulate_relation_uses(&self, relation_counts: &mut HashMap<&'static str, u64>) {
         let Self {
             add,
             add_small,
