+++ stwo_cairo_prover/crates/cairo-air/src/pedersen/air.rs
@@ -27,7 +27,7 @@ impl PedersenContextClaim {
             .unwrap_or_default()
     }
 
-    pub fn accumulate_relation_uses(&self, relation_counts: &mut HashMap<&'static str, u32>) {
+    pub fn accumulate_relation_uses(&self, relation_counts: &mut HashMap<&'static str, u64>) {
         if let Some(claim) = &self.claim {
             claim.accumulate_relation_uses(relation_counts);
         }
@@ -55,7 +55,7 @@ impl Claim {
         TreeVec::concat_cols(log_sizes)
     }
 
-    pub fn accumulate_relation_uses(&self, relation_counts: &mut HashMap<&'static str, u32>) {
+    pub fn accumulate_relation_uses(&self, relation_counts: &mut HashMap<&'static str, u64>) {
         let Self {
             partial_ec_mul,
             pedersen_points_table: _,
